+++ src/main/java/muzusi/application/auth/service/AuthService.java
@@ -8,8 +8,7 @@ import muzusi.application.auth.dto.TokenDto;
 import muzusi.application.auth.dto.UserInfoDto;
 import muzusi.application.auth.dto.UserStatusDto;
 import muzusi.application.auth.helper.JwtHelper;
-import muzusi.application.auth.service.client.OAuthClient;
-import muzusi.application.auth.service.client.OAuthClientFactory;
+import muzusi.infrastructure.client.OAuthClient;
 import muzusi.domain.user.entity.User;
 import muzusi.domain.user.exception.UserErrorType;
 import muzusi.domain.user.service.UserService;
